TransformerAbstract disassemble to traits
